% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{PD_check_missingness}
\alias{PD_check_missingness}
\title{PD_check_missingness}
\usage{
PD_check_missingness(dat, startsWith, prop_good = 0.5, llod = 1e+06)
}
\arguments{
\item{dat}{data.frame to check}

\item{startsWith}{Character value identifying columns to check (passed to \code{starts_with()})}

\item{prop_good}{Numeric proportion of good values to keep}

\item{llod}{Numeric lower limit of detection}
}
\value{
data frame with missingness information and which rows to drop
}
\description{
Check for missingness in data by condition and flag conditions with too many missing or low values.
}
