% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{PD_clean}
\alias{PD_clean}
\title{PD_clean}
\usage{
PD_clean(file, colTypes)
}
\arguments{
\item{file}{file to read in from PD and clean}

\item{colTypes}{column types data.frame from \code{PD_colTypes()}}
}
\value{
data frame with cleaned data
}
\description{
Clean up data
}
