% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{PD_colTypes}
\alias{PD_colTypes}
\title{PD_colTypes}
\usage{
PD_colTypes(colDescriptions)
}
\arguments{
\item{colDescriptions}{list of column descriptions from PD node_args}
}
\value{
tibble with column names and types
}
\description{
This function reads in all column names from the node_args.json file and returns a tibble with the following columns:
\itemize{
\item name: original column name
\item type: expected column type
\item new_name: new column name
}
}
\details{
Fetch column types from PD node_args json file
}
