% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{PD_filter}
\alias{PD_filter}
\title{PD_filter}
\usage{
PD_filter(dat, dat_check, startsWith, ratio)
}
\arguments{
\item{dat}{data.frame to filter}

\item{dat_check}{data.frame with missingness information from \code{PD_check_missingness()}}

\item{startsWith}{Character value identifying columns to check (passed to \code{starts_with()})}

\item{ratio}{Logical value indicating whether the columns to be filtered are ratios}
}
\value{
data frame with filtered data
}
\description{
Filter data by condition
}
