% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_sample_node_args}
\alias{read_sample_node_args}
\title{read_sample_node_args}
\usage{
read_sample_node_args(set = NULL)
}
\arguments{
\item{set}{Numeric value, indicating the sample node_args file to read in}
}
\value{
A list containing the sample node arguments
}
\description{
Read in a sample node arguments file provided with the package
}
