% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{write_PD_response}
\alias{write_PD_response}
\title{write_PD_response}
\usage{
write_PD_response(node_args, dat_filtered, colTypes, startsWith)
}
\arguments{
\item{node_args}{node args from PD}

\item{dat_filtered}{filtered data frame to write out}

\item{colTypes}{column types data.frame from \code{PD_colTypes()}}

\item{startsWith}{Character value identifying columns to check (passed to \code{starts_with()})}
}
\value{
Invisibly returns node_response JSON list
}
\description{
This function writes out the node response file to the location defined in \code{node_args$ExpectedResponsePath}
}
\details{
Write PD response file
}
